feat(sql): Refactor
